// src/app/core/models/film.model.ts
export interface Leinwand {
  leinwand_name: string;
  leinwand_highlights_css: string;
  // vorstellungen: {
  //     belegung_ampel: string;
  //     belegung_ampel_text: string;
  //     tag_der_woche: string;
  //     uhrzeit: string;
  //     datum_uhrzeit: string;
  //     datum_uhrzeit_iso?: string;
  //     deaktiviert?: boolean;
  //     vorstellung_id?: string;
  //     href?: string;
  //     ausgegraut?: boolean;
  // }[];
  vorstellungen: Vorstellung[];
  theater_name: string;
  theater_ort_name: string;
  theater_ort_adresse: string;
  release_flags: any[];
  vorstellungen_anzahl_tage: Record<string, number>;
  hat_aktive_vorstellungen: boolean;
}

export interface Vorstellung {
  belegung_ampel: string;
  belegung_ampel_text: string;
  tag_der_woche: string;
  uhrzeit: string;
  datum_uhrzeit: string;
  datum_uhrzeit_iso?: string;
  deaktiviert?: boolean;
  vorstellung_id?: string;
  href?: string;
  ausgegraut?: boolean;
}

export interface Theater {
  theater_name: string;
  theater_system_id: string;
  theater_selektiert_css: string;
  theater_system_sortierung: string;
  theater_ort_name: string;
  theater_ort_adresse: string;
  leinwaende: Leinwand[];
  vorstellungen_anzahl_tage: Record<string, number>;
}

export interface Film {
  system_id: string;
  film_hauptfilm: string;
  film_ist_ov: string;
  film_neu?: string;
  film_titel: string;
  film_beschreibung: string;
  film_wochentage: {
    tag_der_woche: string;
    tag_text: string;
    tag_datum: string;
  }[];
  theater: Theater[];
  film_system_id: string;
  filminfo_href: string;
  film_cover_src: string;
  film_vorverkauf_zeit_nicht_erreicht?: string;
  vorstellungen_anzahl_tage: Record<string, number>;
  vorstellungen_anzahl_tage_max?: number;

  system_erstell_datum: string;
  system_bearbeit_datum: string;
  film_teasertext: string;
  film_synopsis: string;
  film_dauer: string;
  film_schlagzeile: string;
  film_website: string;
  film_bundesstart_datum: string;
  film_produktionslaender: string;
  film_jahr: string;
  film_centerstart_zeit: string;
  film_edis: string;
  film_fsk: string;
  film_darsteller: string;
  film_regisseure?: string;
  film_produzenten: string;
  film_komponisten: string;
  film_autoren: string;
  film_studios: string;
  film_publisher: string;
  film_daten_sprache: string;
  film_original_sprache: string;
  film_favored: string;
  film_bundesstart_datum_iso: string;
  film_centerstart_zeit_iso: string;
  film_vorverkauf_datum: string;
  film_tags: Tags[];
  barrierefrei_tags?: {
    id: number;
    text: string;
    href: string;
    tags: Tags[];
  };
  main_film_sprache: string;
  film_daten_sprache_text: string;
  film_pressezitate: {
    id: number;
    text: string;
    quelle: string;
  }[];
  trailerUrl?: string;
  trailerPreviewUrl?: string;
}

interface Tags {
  id: number;
  text: string;
  href: string;
}